Gil - splitted the filters class to 3 classes, each representing a type of target filter
