+++ src/main/java/frc/robot/poseestimator/observations/GameObjectPoseObservation.java
@@ -1,0 +1,5 @@
+package frc.robot.poseestimator.observations;
+
+import edu.wpi.first.math.geometry.Pose3d;
+
+public record GameObjectPoseObservation(Pose3d pose, double timestamp) {}
