+++ src/main/java/frc/robot/poseestimator/photonvision/PhotonTargetRawData.java
@@ -3,7 +3,7 @@ package frc.robot.poseestimator.photonvision;
 import edu.wpi.first.math.geometry.Pose3d;
 
 public record PhotonTargetRawData(
-	Pose3d robotPose,
+	Pose3d targetPose,
 	PhotonVisionTarget target,
 	double timestamp,
 	double ambiguity,
